refactor MWPW-120007
