+++ head.html
@@ -1,5 +1,6 @@
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
-<script src="https://cdn.jsdelivr.net/npm/bowser@latest/es5.js"></script>
+<!-- convert to an imported module -->
+<script src="/scripts/bowser.js"></script>
 <script src="/scripts/scripts.js" type="module"></script>
 <link rel="stylesheet" href="/styles/styles.css"/>
 <style>body { display: none; }</style>
